KimEuiRyeong/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
arelle-release==2.37.34
attrs==25.3.0
beautifulsoup4==4.13.4
bottle==0.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
dart-fss==0.4.13
distro==1.9.0
et_xmlfile==2.0.0
fake-useragent==2.2.0
filelock==3.18.0
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
isodate==0.7.2
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain-core==0.3.66
langchain-openai==0.3.27
langchain-pinecone==0.2.8
langchain-tests==0.3.20
langchain-text-splitters==0.3.8
langsmith==0.4.4
lxml==5.4.0
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
numpy==2.3.1
openai==1.93.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.3.0
pinecone==7.3.0
pinecone-plugin-assistant==1.7.0
pinecone-plugin-interface==0.0.7
pluggy==1.6.0
propcache==0.3.2
py-cpuinfo==9.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyMuPDF==1.26.3
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==0.26.0
pytest-benchmark==5.1.0
pytest-codspeed==3.2.0
pytest-recording==0.13.4
pytest-socket==0.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
syrupy==4.9.1
tenacity==9.1.2
termcolor==2.3.0
tiktoken==0.9.0
tqdm==4.67.1
truststore==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
vcrpy==7.0.0
wrapt==1.17.2
xmltodict==0.14.2
yarl==1.20.1
yaspin==3.1.0
zstandard==0.23.0
